pyproject.toml
[project]
name = "my_masters_degree"
version = "0.1.0"
description = "A Python project for my masters degree research"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.5.0",
    "google-genai>=1.57.0",
    "matplotlib>=3.10.8",
    "numpy>=2.3.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pandera>=0.29.0",
    "pdfplumber>=0.11.9",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "scipy>=1.17.0",
    "torch>=2.9.1",
    "torchcodec>=0.9.1",
    "werpy>=3.1.0",
]
authors = [{ name = "Antonio", email = "antonionmoreira@gmail.com" }]
keywords = ["research", "masters", "thesis"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pyrefly]
project-includes = ["**/*.py*", "**/*.ipynb"]

[dependency-groups]
dev = [
    "jupyterlab>=4.4.7",
    "pip>=26.0.1",
    "setuptools>=80.9.0",
    "wheel>=0.45.1",
]
